+++ src/app/models/ambito.ts
@@ -1,0 +1,5 @@
+export interface Ambito {
+  id_ambito: number;
+  nombre_ambito: string;
+  activo: number;
+}