+++ src/app/models/componente.ts
@@ -8,4 +8,5 @@ export interface Componente {
     activo: number;
     id_proyecto?: number; // Mantenemos el ID por si acaso
     proyecto?: Proyecto;    // Se añade el objeto Proyecto
+     id_ambito: number;
 }