+++ src/app/pages/auth/login.ts
@@ -53,13 +53,13 @@ import { MessageService } from 'primeng/api';
                                 <label for="password" class="block text-surface-900 dark:text-surface-0 font-medium text-xl mb-2">Password</label>
                                 <p-password id="password" placeholder="Password" [toggleMask]="true" styleClass="mb-4" [fluid]="true" [feedback]="false" formControlName="password"></p-password>
 
-                                <div class="flex items-center justify-between mt-2 mb-8 gap-8">
+                                <!-- <div class="flex items-center justify-between mt-2 mb-8 gap-8">
                                     <div class="flex items-center">
-                                        <p-checkbox id="rememberme1" binary class="mr-2"></p-checkbox>
+                                         <p-checkbox id="rememberme1" binary class="mr-2"></p-checkbox>
                                         <label for="rememberme1">Recuérdame</label>
                                     </div>
                                     <span class="font-medium no-underline ml-2 text-right cursor-pointer text-primary">¿Olvidó su contraseña?</span>
-                                </div>
+                                </div> -->
 
 
                                 <p-button label="Entrar" styleClass="w-full" type="submit" [disabled]="loginForm.invalid"></p-button>