+++ CHANGELOG.md
@@ -3,10 +3,17 @@ Todos los cambios significativos en este proyecto se documentarán en este archi
 El formato se basa en [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 y este proyecto adhiere al [Versionado Semántico](https://semver.org/lang/es/).
 
+## [0.4.1] - 2025-08-18
+
+### Added
+-   Se agregó un campo para compontente: ámbito.
+-   Se agregó a la configuracion el campo ámbito.
+
 ## [0.4.0] - 2025-08-14
 
 ### Added
 -   Se agregaron las configuraciones para estados, proyectos y criticidades.
+-   Se agregaron opciones de importar y exportar en casos
 
 
 ## [0.3.1] - 2025-08-14